Styling fixes and StackedBarChart UI refactoring
