+++ src/components/Graphs/index.ts
@@ -1,6 +1,6 @@
 import LineChart from "./LineChart/LineChart";
 import BarChart from "./BarChart/BarChart";
-import StackedBarChart from "./StackedBarChart";
+import StackedBarChart from "./StackedBarChart/StackedBarChart";
 import SmallDataBox from "./SmallDataBox/SmallDataBox";
 import WeeklyProgress from "./WeeklyProgress/WeeklyProgress";
 import BooleanBox from "./BooleanBox/BooleanBox";
